requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
google-genai==1.14.0
python-dotenv==1.1.0
jinja2==3.1.6
